requirements.txt
python-telegram-bot[job-queue]
requests
schedule
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
pytz
python-dotenv
